[lua --> C++]

+++ cocos2d-x-4.0/tests/cpp-empty-test/Classes/HelloWorldScene.cpp
@@ -69,7 +69,7 @@ bool HelloWorld::init()
    // add a label shows "Hello World"
    // create and initialize a label
 
-   auto label = Label::createWithTTF("Hello World", "fonts/arial.ttf", TITLE_FONT_SIZE);
+   auto label = Label::createWithTTF("Hello World", "fonts/arial.ttf", (float) TITLE_FONT_SIZE);
 
    // position the label on the center of the screen
     label->setPosition(origin.x + visibleSize.width/2,
